Улучшена структура
