+++ src/BitrixDebugger/Debugger/Debugger.php
@@ -12,16 +12,6 @@ use GGrach\BitrixDebugger\Configurator\DebugBarConfigurator;
  */
 class Debugger extends LogFileDebugger {
 
-    /**
-     * @var DebuggerConfigurator 
-     */
-    protected $configuratorDebugger;
-
-    /**
-     * @var DebugBarConfigurator 
-     */
-    protected $configuratorDebugBar;
-
     public function __construct($debuggerConfigurator = null, $debugBarConfigurator = null) {
         if ($debuggerConfigurator === null) {
             $this->setConfiguratorDebugger(new DebuggerConfigurator());
@@ -34,22 +24,6 @@ class Debugger extends LogFileDebugger {
         } elseif ($debugBarConfigurator instanceof DebugBarConfigurator) {
             $this->setConfiguratorDebugBar($debugBarConfigurator);
         }
-    }
-
-    public function getConfiguratorDebugger(): DebuggerConfigurator {
-        return $this->configuratorDebugger;
-    }
-
-    public function getConfiguratorDebugBar(): DebugBarConfigurator {
-        return $this->configuratorDebugBar;
-    }
-
-    public function setConfiguratorDebugger(DebuggerConfigurator $configuratorDebugger): void {
-        $this->configuratorDebugger = $configuratorDebugger;
-    }
-
-    public function setConfiguratorDebugBar(DebugBarConfigurator $configuratorDebugBar): void {
-        $this->configuratorDebugBar = $configuratorDebugBar;
     }
 
 }
